Backend/productDomain/product_create_service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymssql
python-dotenv
python-multipart